Reintroduced the currently missing --list function
And renamed the examples directory to the templates directory. Shipped out of the box they may all be examples, but they are after all, templates. Also renamed the mandatory .html filename of a template from index.html back to template.html because this legacy practice provides no advantage in this environment and is semantically less acurate of the function.